refactor(routing): extract useRouteTransition composable and route-skeleton
