+++ src/locales/en-us.json
@@ -1,5 +1,4 @@
 {
-  "nav.back": "Back",
   "ui-kit.alert.cancel": "Cancel",
   "ui-kit.alert.continue": "Continue",
   "app.title": "TaroFlash",
